+++ app/src/main/java/com/fibelatti/photowidget/home/MyWidgetsScreen.kt
@@ -105,8 +105,7 @@ fun MyWidgetsScreen(
                                 opacity = widget.opacity,
                                 modifier = Modifier.fillMaxSize(),
                                 blackAndWhite = widget.blackAndWhite,
-                                borderColorHex = widget.borderColor,
-                                borderWidth = widget.borderWidth,
+                                border = widget.border,
                                 isLoading = widget.isLoading,
                             )